refactor: split app coordinator services
