+++ src/hooks/useAppCoordinatorAuthLoaders.ts
@@ -1,0 +1,66 @@
+import { useAppAuthActions } from './useAppAuthActions';
+import { useAppTabDataLoaders } from './useAppTabDataLoaders';
+import type { CoordinatorServicesArgs } from './useAppCoordinatorServices.types';
+
+export function useAppCoordinatorAuthLoaders({
+  routeState,
+  domainState,
+  dataState,
+}: CoordinatorServicesArgs) {
+  const { activeTab } = routeState;
+  const {
+    auth: { sessionUser },
+  } = domainState;
+  const {
+    adminSummary,
+    communityRoutesCacheRef,
+    coursesLoadedRef,
+    feedLoadedRef,
+    myPage,
+    replaceCommunityRoutes,
+    setAdminLoading,
+    setAdminSummary,
+    setCommunityRoutes,
+    setCourses,
+    setMyPage,
+    setReviews,
+  } = dataState;
+
+  const { startProviderLogin, handleUpdateProfile, handleLogout } = useAppAuthActions({
+    setMyPage,
+    formatErrorMessage,
+  });
+
+  const dataLoaders = useAppTabDataLoaders({
+    activeTab,
+    adminSummary,
+    myPage,
+    sessionUser,
+    communityRoutesCacheRef,
+    feedLoadedRef,
+    coursesLoadedRef,
+    replaceCommunityRoutes,
+    setCommunityRoutes,
+    setReviews,
+    setCourses,
+    setAdminLoading,
+    setAdminSummary,
+    setMyPage,
+  });
+
+  return {
+    sessionUser,
+    myPage,
+    startProviderLogin,
+    handleUpdateProfile,
+    handleLogout,
+    dataLoaders,
+  };
+}
+
+function formatErrorMessage(error: unknown) {
+  if (error instanceof Error) {
+    return error.message;
+  }
+  return '?遺욧퍕??筌ｌ꼶???? 筌륁궢六??곸뒄. ?醫롫뻻 ??쇰퓠 ??쇰뻻 ??뺣즲??雅뚯눘苑??';
+}
