+++ src/hooks/useAppCoordinatorServices.ts
@@ -1,25 +1,7 @@
-import { useAppAuthActions } from './useAppAuthActions';
-import { useAppTabDataLoaders } from './useAppTabDataLoaders';
-import { useAppNavigationHelpers } from './useAppNavigationHelpers';
-import { useGlobalNotifications } from './useGlobalNotifications';
-import { useAppViewModels } from './useAppViewModels';
-import { useAppPagePaginationActions } from './useAppPagePaginationActions';
-import { useActiveReviewComments } from './useActiveReviewComments';
-import type {
-  DataState,
-  DomainState,
-  PageRuntimeState,
-  RouteState,
-  ShellRuntimeState,
-} from './useAppShellCoordinator.types';
-
-type CoordinatorServicesArgs = {
-  routeState: RouteState;
-  domainState: DomainState;
-  shellRuntimeState: ShellRuntimeState;
-  pageRuntimeState: PageRuntimeState;
-  dataState: DataState;
-};
+import { useAppCoordinatorAuthLoaders } from './useAppCoordinatorAuthLoaders';
+import { useAppCoordinatorNavigationNotifications } from './useAppCoordinatorNavigationNotifications';
+import { useAppCoordinatorViewState } from './useAppCoordinatorViewState';
+import type { CoordinatorServicesArgs } from './useAppCoordinatorServices.types';
 
 export function useAppCoordinatorServices({
   routeState,
@@ -28,186 +10,34 @@ export function useAppCoordinatorServices({
   pageRuntimeState,
   dataState,
 }: CoordinatorServicesArgs) {
-  const {
-    activeTab,
-    drawerState,
-    selectedPlaceId,
-    selectedFestivalId,
-    commitRouteState,
-    goToTab,
-  } = routeState;
-  const {
-    auth: { sessionUser },
-    map: { activeCategory, selectedRoutePreview, setSelectedRoutePreview },
-    myPage: { myPageTab, setMyPageTab },
-    returnView: { setReturnView },
-    review: {
-      feedPlaceFilterId,
-      activeCommentReviewId,
-      highlightedCommentId,
-      highlightedReviewId,
-      setActiveCommentReviewId,
-      setFeedPlaceFilterId,
-      setHighlightedCommentId,
-      setHighlightedReviewId,
-      setHighlightedRouteId,
-    },
-  } = domainState;
-  const {
-    notice,
-    setNotice,
-    currentPosition,
-    mapLocationStatus,
-    mapLocationMessage,
-    bootstrapStatus,
-    bootstrapError,
-  } = shellRuntimeState;
-  const {
-    adminSummary,
-    communityRoutesCacheRef,
-    coursesLoadedRef,
-    feedLoadedRef,
-    festivals,
-    myPage,
-    places,
-    replaceCommunityRoutes,
-    reviews,
-    selectedPlaceReviews,
-    setAdminLoading,
-    setAdminSummary,
-    setCommunityRoutes,
-    setCourses,
-    setMyPage,
-    setReviews,
-    stampState,
-  } = dataState;
-
-  const { startProviderLogin, handleUpdateProfile, handleLogout } = useAppAuthActions({
-    setMyPage,
-    formatErrorMessage,
+  const authLoaders = useAppCoordinatorAuthLoaders({
+    routeState,
+    domainState,
+    shellRuntimeState,
+    pageRuntimeState,
+    dataState,
   });
 
-  const dataLoaders = useAppTabDataLoaders({
-    activeTab,
-    adminSummary,
-    myPage,
-    sessionUser,
-    communityRoutesCacheRef,
-    feedLoadedRef,
-    coursesLoadedRef,
-    replaceCommunityRoutes,
-    setCommunityRoutes,
-    setReviews,
-    setCourses,
-    setAdminLoading,
-    setAdminSummary,
-    setMyPage,
-  });
+  const navigationNotifications = useAppCoordinatorNavigationNotifications({
+    routeState,
+    domainState,
+    shellRuntimeState,
+    pageRuntimeState,
+    dataState,
+  }, authLoaders);
 
-  const navigationHelpers = useAppNavigationHelpers({
-    activeTab,
-    myPageTab,
-    activeCommentReviewId,
-    highlightedCommentId,
-    highlightedReviewId,
-    selectedPlaceId,
-    selectedFestivalId,
-    drawerState,
-    feedPlaceFilterId,
-    reviews,
-    selectedPlaceReviews,
-    myPageReviews: myPage?.reviews ?? [],
-    setActiveCommentReviewId,
-    setHighlightedCommentId,
-    setHighlightedReviewId,
-    setHighlightedRouteId,
-    setReturnView,
-    setSelectedRoutePreview,
-    setFeedPlaceFilterId,
-    setNotice,
-    goToTab,
-    commitRouteState,
-    upsertReviewCollections: dataState.upsertReviewCollections,
-  });
-
-  const {
-    notifications,
-    unreadNotificationCount,
-    handleMarkNotificationRead,
-    handleMarkAllNotificationsRead,
-    handleDeleteNotification,
-    handleOpenGlobalNotification,
-  } = useGlobalNotifications({
-    sessionUser,
-    myPage,
-    goToTab,
-    setMyPageTab,
-    handleOpenCommentWithReturn: navigationHelpers.handleOpenCommentWithReturn,
-    handleOpenReviewWithReturn: navigationHelpers.handleOpenReviewWithReturn,
-  });
-
-  const viewModels = useAppViewModels({
-    places,
-    festivals,
-    reviews,
-    selectedPlaceReviews,
-    selectedPlaceId,
-    selectedFestivalId,
-    selectedRoutePreview,
-    activeCategory,
-    myPage,
-    notifications,
-    unreadNotificationCount,
-    stampState,
-    currentPosition,
-    sessionUser,
-    notice,
-    bootstrapStatus,
-    bootstrapError,
-    mapLocationStatus,
-    mapLocationMessage,
-  });
-
-  const paginationActions = useAppPagePaginationActions({
-    sessionUser,
-    myPage,
-    setReviews,
-    setMyPage,
-    reportBackgroundError,
-  });
-
-  const activeReviewCommentsState = useActiveReviewComments({
-    activeCommentReviewId,
-    setNotice,
-    formatErrorMessage,
-  });
+  const viewState = useAppCoordinatorViewState({
+    routeState,
+    domainState,
+    shellRuntimeState,
+    pageRuntimeState,
+    dataState,
+  }, authLoaders, navigationNotifications);
 
   return {
-    startProviderLogin,
-    handleUpdateProfile,
-    handleLogout,
-    dataLoaders,
-    navigationHelpers,
-    notifications,
-    unreadNotificationCount,
-    handleMarkNotificationRead,
-    handleMarkAllNotificationsRead,
-    handleDeleteNotification,
-    handleOpenGlobalNotification,
-    viewModels,
-    paginationActions,
-    activeReviewCommentsState,
+    ...authLoaders,
+    ...navigationNotifications,
+    ...viewState,
     pageRuntimeState,
   };
-}
-
-function formatErrorMessage(error: unknown) {
-  if (error instanceof Error) {
-    return error.message;
-  }
-  return '?붿껌??泥섎━?섏? 紐삵뻽?댁슂. ?좎떆 ?ㅼ뿉 ?ㅼ떆 ?쒕룄??二쇱꽭??';
-}
-
-function reportBackgroundError(error: unknown) {
-  console.error(error);
 }
