+++ src/hooks/useAppCoordinatorServices.types.ts
@@ -1,0 +1,15 @@
+import type {
+  DataState,
+  DomainState,
+  PageRuntimeState,
+  RouteState,
+  ShellRuntimeState,
+} from './useAppShellCoordinator.types';
+
+export type CoordinatorServicesArgs = {
+  routeState: RouteState;
+  domainState: DomainState;
+  shellRuntimeState: ShellRuntimeState;
+  pageRuntimeState: PageRuntimeState;
+  dataState: DataState;
+};
